Remove deprecated example codes and refactor useful ones
