--- Examples/log.csv
@@ -1,3 +1,0 @@
-0.000000, 0.000000,
-1.000000, 1.000000,
-2.000000, 4.000000,
