--- Examples/sampleData.txt
@@ -1,18 +1,0 @@
-0,3
-1,4
-2,5
-3,8
-4,3
-5,0
-6,1
-7,8
-8,9
-9,12
-10,7
-11,4
-12,6
-13,2
-14,10
-15,11
-16,11
-17,-1
