--- Examples/test_Sound.py
@@ -1,5 +1,0 @@
-import os
-
-print '\007'
-
-os.system('spd-say "your program has finished"')
