+++ Examples/torque_speed_curve.py
